pyproject.toml
[project]
name = "longcon-mcp-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jira>=3.8.0",
    "fastmcp>=0.1.0",
    "python-dotenv>=1.1.1",
    "ipython>=9.5.0",
]
